backend/requirements.txt
# API framework
fastapi
uvicorn[standard]
python-multipart

# Authentication
python-jose[cryptography]

# Configuration
python-dotenv
pydantic
pydantic-settings

# Image processing & OCR
Pillow
pytesseract

# Discogs
python3-discogs-client

# YouTube search
youtube-search-python
google-api-python-client

# SoundCloud
soundcloud-v2

# HTTP client
httpx

selenium